// crates/desktop/src/generated/CopyRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitSourceDto } from "./GitSourceDto";

export type CopyRequest = { "kind": "files", roots: Array<string>, paths: Array<string>, } | { "kind": "git", repo: string, roots: Array<string>, source: GitSourceDto, };